pyproject.toml
[project]
name = "jarvis-mark-xl"
version = "3.0.0"
description = "JARVIS Mark-XL — Local-First Agentic Desktop Assistant"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name = "Rohan"}]

dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "httpx>=0.27.0",
    
    # Data validation
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    
    # LLM SDK
    "google-genai>=1.0.0",
    
    # Memory
    "chromadb>=0.5.0",
    "sentence-transformers>=3.0.0",
    "aiosqlite>=0.20.0",
    
    # System tools
    "psutil>=5.9.0",
    "pyautogui>=0.9.54",
    
    # Screen capture
    "mss>=9.0.0",
    "Pillow>=10.0.0",
]

[project.optional-dependencies]
windows = [
    "pywin32>=306",
]
browser = [
    "playwright>=1.45.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
]

[project.scripts]
jarvis = "backend.main:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py310"
line-length = 100
